+++ com.ibm.streamsx.monitoring/impl/java/src/com/ibm/streamsx/monitoring/jobs/internal/filter/InstanceFilter.java
@@ -5,7 +5,7 @@
 // ****************************************************************************
 //
 
-package com.ibm.streamsx.monitoring.notification.filter;
+package com.ibm.streamsx.monitoring.jobs.internal.filter;
 
 import java.math.BigInteger;
 import java.util.HashMap;
