+++ com.ibm.streamsx.monitoring/impl/java/src/com/ibm/streamsx/monitoring/jobs/internal/filter/OperatorFilter.java
@@ -5,9 +5,8 @@
 // ****************************************************************************
 //
 
-package com.ibm.streamsx.monitoring.notification.filter;
+package com.ibm.streamsx.monitoring.jobs.internal.filter;
 
-import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
@@ -15,17 +14,20 @@ import java.util.regex.PatternSyntaxException;
 
 import org.apache.log4j.Logger;
 
-final class PeFilter {
+final class OperatorFilter extends PatternMatcher {
 
 	/**
 	 * Logger for tracing.
 	 */
-	private static Logger _trace = Logger.getLogger(PeFilter.class.getName());
+	private static Logger _trace = Logger.getLogger(OperatorFilter.class.getName());
 
-
-	public PeFilter() throws PatternSyntaxException {
+	public OperatorFilter(String regularExpression) throws PatternSyntaxException {
+		super(regularExpression);
 	}
 
-
+	public boolean matchesOperatorName(String operatorName) {
+		boolean matches = matches(operatorName);
+		return matches;
+	}
 
 }
