+++ com.ibm.streamsx.monitoring/impl/java/src/com/ibm/streamsx/monitoring/jobs/internal/filter/PeFilter.java
@@ -5,8 +5,9 @@
 // ****************************************************************************
 //
 
-package com.ibm.streamsx.monitoring.notification.filter;
+package com.ibm.streamsx.monitoring.jobs.internal.filter;
 
+import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
@@ -14,20 +15,17 @@ import java.util.regex.PatternSyntaxException;
 
 import org.apache.log4j.Logger;
 
-final class OperatorFilter extends PatternMatcher {
+final class PeFilter {
 
 	/**
 	 * Logger for tracing.
 	 */
-	private static Logger _trace = Logger.getLogger(OperatorFilter.class.getName());
+	private static Logger _trace = Logger.getLogger(PeFilter.class.getName());
 
-	public OperatorFilter(String regularExpression) throws PatternSyntaxException {
-		super(regularExpression);
+
+	public PeFilter() throws PatternSyntaxException {
 	}
 
-	public boolean matchesOperatorName(String operatorName) {
-		boolean matches = matches(operatorName);
-		return matches;
-	}
+
 
 }
